More logical function renaming and restructuring
