+++ esn_dev/optimize.py
@@ -2,16 +2,37 @@ import numpy as np
 import gc #garbage collector
 import scipy
 
-def lstsq_stable(H, labels, thresh=1e-4):
-    """    if labels.ndim != 2:
-        raise ValueError("Labels must have shape (time, features)")
+def create_readout_matrix(dynsys, H, targets):
+    """
+    Compute the output matrix `Who` via least squares. 
+    and add it to the dynsys tuple to create a model tuple.
+
+    Params:
+        dynsys: ESN tuple (mapih, Whh, bh)
+        H:      state matrix as generated by 
+                `evolve_hidden_state` in mode `train`.
+        targets: training targets (time, height, width)
+        
+    Returns:
+        
+        (mapih, Whh, bh, Who): dynsys + Who tuple
+    
+    """
+    
+    Who = lstsq_stable(states, targets)
+    
+    return dynsys + (Who,)
+
+def lstsq_svd(H, targets, thresh=1e-4):
+    """    if targets.ndim != 2:
+        raise ValueError("Targets must have shape (time, features)")
     
     print('Doing svd')
     U, s, Vh = scipy.linalg.svd(H.T, full_matrices=False)
     scale = s[0]
     n = np.sum(np.abs(s/scale) > thresh)  # Ensure condition number less than 1/thresh
     
-    L = labels.T
+    L = targets.T
     v = Vh[:n, :].T
     uh = U[:, :n].T
     
@@ -27,15 +48,15 @@ def lstsq_stable(H, labels, thresh=1e-4):
     #print(f'Tikhonov with parameter {alpha}, largest Singualar value: {s[0]:.3e}')
     #Who =  np.dot(s*np.dot(L, Vh.T) / (alpha +s**2), U.T)"""
     
-    if labels.ndim != 2:
-        raise ValueError("Labels must have shape (time, features)")
+    if targets.ndim != 2:
+        raise ValueError("targets must have shape (time, features)")
     
     print('Doing svd')
     U, s, Vh = scipy.linalg.svd(H, full_matrices=False)
     scale = s[0]
     n = np.sum(np.abs(s/scale) > thresh)  # Ensure condition number less than 1/thresh
     
-    L = labels
+    L = targets
     v = Vh[:n, :]
     uh = U[:, :n]
     
@@ -46,7 +67,7 @@ def lstsq_stable(H, labels, thresh=1e-4):
     del H
     gc.collect()
     print('deleted H, U')
-    print(f'Labels have shape {L.shape}')
+    print(f'targets have shape {L.shape}')
     Who =  np.dot(np.dot(L.T, uh) / s[:n], v) 
     #alpha = s[0]*0.001
     #print(f'Tikhonov with parameter {alpha}, largest Singualar value: {s[0]:.3e}')
@@ -54,11 +75,11 @@ def lstsq_stable(H, labels, thresh=1e-4):
 
     return Who
 
-def lstsq_pcr(H, labels, thresh=1e-4):
+def lstsq_pcr(H, targets, thresh=1e-4):
     global Vh
 
-    if labels.ndim != 2:
-        raise ValueError("Labels must have shape (time, features)")
+    if targets.ndim != 2:
+        raise ValueError("targets must have shape (time, features)")
     print('Doing svd')
     U, s, Vh = scipy.linalg.svd(H, full_matrices=False)
     scale = s[0]
@@ -75,7 +96,7 @@ def lstsq_pcr(H, labels, thresh=1e-4):
     scale = s[0]
     n = np.sum(np.abs(s/scale) > thresh)  # Ensure condition number less than 1/thresh
     
-    L = labels.T
+    L = targets.T
     v = Vh[:n, :].T
     uh = U[:, :n].T  
 
@@ -84,34 +105,4 @@ def lstsq_pcr(H, labels, thresh=1e-4):
     #Who =  np.dot(np.dot(L, v) / s[:n], uh) 
     Who =  np.dot(np.dot(L, v) / s[:n], uh) 
    
-    return Who
-
-
-
-
-def imed_lstsq_stable(states, inputs, labels, sigma):
-    flat_inputs = inputs.reshape(inputs.shape[0], -1)
-    flat_labels = labels.reshape(inputs.shape[0], -1)
-
-    # prep IMED
-    G = imed_matrix(inputs.shape[1:], sigma=sigma)
-    (w,V) = np.linalg.eigh(G)
-
-    #thresh = 0
-    #n = np.sum(np.abs(w) < thresh)
-    #s = np.sqrt(w[n:])
-    #V = V[n:,:]
-    s = np.sqrt(w)
-
-    G12 = np.dot(V.T, s[:,None]*V)
-
-    # transform inputs / labels
-    flat_inputs = np.matmul(G12, flat_inputs[:,:,None])[:,:,0]
-    flat_labels = np.matmul(G12, flat_labels[:,:,None])[:,:,0]
-
-    # compute Who
-    Who  = lstsq_stable(states, flat_labels)
-    s    = 1/s
-    iG12 = np.dot(V.T,s[:,None]*V)
-    Who  = np.dot(iG12, Who)
     return Who
